get doze not complaining about bp::file::size()
git-svn-id: svn+ssh://corpsvnproxy/yahoo/BrowserPlus/public_platform/trunk@348 3fcf768e-b076-0410-b9b8-e8c34e2d470b

+++ src/libs/bputils/src/Http/Client/HttpTransaction_Windows.cpp
@@ -200,10 +200,10 @@ private:
 
     // progress info
     IListener*      m_pListener;
-    uintmax_t       m_sendTotalBytes;
-    uintmax_t       m_bytesSent;
-    uintmax_t       m_receiveTotalBytes;
-    uintmax_t       m_bytesReceived;
+    size_t          m_sendTotalBytes;
+    size_t          m_bytesSent;
+    size_t          m_receiveTotalBytes;
+    size_t          m_bytesReceived;
     bool            m_zeroProgressSent;
     double          m_lastProgressSent;
 
@@ -906,7 +906,7 @@ Transaction::Impl::sendRequestWithBody()
             return GetLastError();
         }
         m_sendTotalBytes = bp::file::size(path);
-    } else {
+	} else {
         m_ePostSource = eFromBuffer;
         m_sendTotalBytes = m_pRequest->body.size();
     }
